requirements.txt
graphql-core==0.5.1
Cython>=0.24
venusian>=1.0
pytest>=2.9.1
memoized-property==1.0.2
marshmallow==2.7.3
